Added recommended changes
