+++ tests/fixtures/datasets.py
@@ -5,7 +5,7 @@ import numpy as np
 import pandas as pd
 import pytest
 import sqlalchemy as sa
-from sklearn.datasets import load_boston, load_iris, fetch_openml
+from sklearn.datasets import load_boston, load_iris
 
 from ml_tooling.data import SQLDataset, FileDataset
 from ml_tooling.data.base_data import Dataset
@@ -180,21 +180,3 @@ def iris_filedataset():
             return df.drop(columns="target").iloc[[idx]]
 
     return IrisFileDataset
-
-
-@pytest.fixture()
-def australian_dataset():
-    australian_data = fetch_openml("Australian", as_frame=True)
-    australian_data_df = pd.DataFrame(
-        data=australian_data.data, columns=australian_data.feature_names
-    )
-    australian_data_df_target = australian_data.target.astype("int")
-
-    class AustralianData(Dataset):
-        def load_training_data(self):
-            return australian_data_df, australian_data_df_target
-
-        def load_prediction_data(self):
-            return australian_data_df
-
-    return AustralianData()
